+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/Android/DistributeInternal.cs
@@ -4,6 +4,7 @@
 #if UNITY_ANDROID && !UNITY_EDITOR
 using System;
 using System.Collections.Generic;
+using Microsoft.AppCenter.Unity.Internal;
 using UnityEngine;
 
 namespace Microsoft.AppCenter.Unity.Distribute.Internal